requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.94.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
pytz==2023.3.post1
setuptools==69.0.2
sqlparse==0.4.4
tzdata==2023.3
